+++ modules/bypass_forbidden.py
@@ -50,9 +50,9 @@ def IP_authorization(s, res, url, domain, page, exclude_len):
 	
 	try:
 		website_ip = socket.gethostbyname(domain)
-		ips_type  = [website_ip, "127.0.0.1", "*", "8.8.8.8", "null", "192.168.0.2", "10.0.0.1", "0.0.0.0","::1","0:0:0:0:0:0:0:1"]
+		ips_type  = [website_ip, "127.0.0.1", "127.0.0.2", "*", "8.8.8.8", "null", "192.168.0.2", "10.0.0.1", "0.0.0.0","::1","0:0:0:0:0:0:0:1"]
 	except:
-		ips_type  = ["127.0.0.1", "*", "8.8.8.8", "null", "192.168.0.2", "10.0.0.1", "localhost", "0.0.0.0","::1","0:0:0:0:0:0:0:1"]
+		ips_type  = ["127.0.0.1", "127.0.0.2", "*", "8.8.8.8", "null", "192.168.0.2", "10.0.0.1", "localhost", "0.0.0.0","::1","0:0:0:0:0:0:0:1"]
 	for h in headers_type:
 		for ip in ips_type:
 			headers = {h : ip}